moved fetchFolderNames to helper utils
