+++ client/src/components/File.tsx
@@ -20,7 +20,7 @@ import ImageIcon from '@mui/icons-material/Image';
 import MusicNoteIcon from '@mui/icons-material/MusicNote';
 import MovieIcon from '@mui/icons-material/Movie';
 import TableChartIcon from '@mui/icons-material/TableChart';
-import { getUsernameById, downloadFile } from '../helper/helperRequests';
+import { getUsernameById, downloadFile } from '../utils/helperRequests';
 import RenameFileDialog from './RenameDialog';
 import FavoriteIcon from '@mui/icons-material/Favorite';
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
