+++ client/src/utils/helperRequests.ts
@@ -37,3 +37,24 @@ export async function downloadFile(
     console.error('Download failed', error);
   }
 }
+
+export async function fetchFolderNames(
+  folderIds: string[],
+): Promise<{ [key: string]: string }> {
+  try {
+    const nameRequests = folderIds.map((id) =>
+      axios.get(`http://localhost:5001/api/folder/foldername/${id}`),
+    );
+    const nameResponses = await Promise.all(nameRequests);
+
+    const folderNames: { [key: string]: string } = {};
+    folderIds.forEach((id, index) => {
+      folderNames[id] = nameResponses[index].data;
+    });
+
+    return folderNames;
+  } catch (error) {
+    console.error('Error fetching folder names:', error);
+    return {};
+  }
+}
